bg_remover/requirements.txt
# Background Remover Pro - Dependencies
# =====================================

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.0

# AI Background Removal
rembg>=2.0.50
onnxruntime>=1.16.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# AI Image Upscaling (Real-ESRGAN)
realesrgan>=0.3.0
torch>=2.0.0
torchvision>=0.15.0
basicsr>=1.4.2
gfpgan>=1.3.8

# Utils
python-dotenv>=1.0.0
watchdog>=3.0.0
